shift main function to a separate file
